refactor(LottoProperty): lotto.constant 로 패키지 변경
[세부사항]
- LottoProperty 상수들이 domain 뿐만 아니라 전체적으로 쓰이기 때문에 변경
